// itisolution-couriers-main/v1/src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface HomePage {
  id: number;
  hero: {
    /**
     * Nagłówek głównej sekcji strony.
     */
    title: string;
    /**
     * Krótki opis sekcji hero.
     */
    description: string;
    /**
     * Frazy wyświetlane rotacyjnie.
     */
    phrases?:
      | {
          phrase: string;
          id?: string | null;
        }[]
      | null;
  };
  functionalities: {
    title: string;
    description: string;
    items?:
      | {
          title: string;
          description: string;
          image: number | Media;
          id?: string | null;
        }[]
      | null;
  };
  pricing: {
    title: string;
    description: string;
    items?:
      | {
          title: string;
          description: string;
          price: string;
          advantages?:
            | {
                advantage: string;
                id?: string | null;
              }[]
            | null;
          recommended: boolean;
          id?: string | null;
        }[]
      | null;
  };
  opinions: {
    title: string;
    description: string;
    items?:
      | {
          opinion: string;
          author: string;
          role: string;
          avatar: number | Media;
          id?: string | null;
        }[]
      | null;
  };
  faq: {
    title: string;
    description: string;
    items?:
      | {
          question: string;
          answer: string;
          id?: string | null;
        }[]
      | null;
  };
  contact: {
    title: string;
    description: string;
    features?:
      | {
          /**
           * Wybierz ikonę
           */
          icon: string;
          title: string;
          description: string;
          id?: string | null;
        }[]
      | null;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}